refactor: App.js内のSpotify SDK初期化の副作用をカスタムフックに切り出し
