+++ src/App.js
@@ -4,13 +4,13 @@ import { BrowserRouter } from "react-router-dom";
 import { SearchProvider } from "./contexts/SearchContext";
 import usePlayerStore from "./store/playerStore";
 import useInitSpotifyToken from "./hooks/useInitSpotifyToken";
+import useInitSpotifyPlayer from "./hooks/useInitSpotifyPlayer";
 
 import Header from "./components/Header";
 import Main from "./components/Main";
 
 function App() {
   const [profile, setProfile] = useState(null);
-  const initPlayer = usePlayerStore((state) => state.initPlayer);
   const player = usePlayerStore((state) => state.player);
   const isSpotifyPlaying = usePlayerStore((state) => state.isSpotifyPlaying);
   const isLocalPlaying = usePlayerStore((state) => state.isLocalPlaying);
@@ -19,17 +19,7 @@ function App() {
 
   useInitSpotifyToken();
 
-  useEffect(() => {
-    let instance;
-
-    (async () => {
-      instance = await initPlayer();
-    })();
-
-    return () => {
-      if (instance?.disconnect) instance.disconnect();
-    };
-  }, []);
+  useInitSpotifyPlayer();
 
   useEffect(() => {
     const cleanup = syncSpotifyPlayerState();
